Revert "Merge branch 'master' of https://github.com/maigiordano28/ddaObli"
This reverts commit 7d8374d46949de0cb1e7edb507b2c8ed02101d15, reversing
changes made to bd59912983d3b400e5cfb7a0f8a35722abbfbd8e.

+++ ObligatorioDDA/src/Dominio/Fachada.java
@@ -39,7 +39,7 @@ public class Fachada extends observable {
         return sUsuario.LoginAdmin(nombreCompleto,password);
     }
 
-    public void agregarJugador(String cedula, String password, String nombre, Double saldo) {
+    public void agregarJugador(String cedula, String password, String nombre, int saldo) {
         sUsuario.agregarJugador(nombre,cedula,password,saldo);
        
     }
